final changes made as per evaluation

+++ FusionIIIT/applications/office_module/views.py
@@ -865,19 +865,19 @@ def apply_purchase(request):
 #    user = get_object_or_404(User,username=request.user.username)
 #    user=ExtraInfo.objects.get(user=user)
     current_user = get_object_or_404(User, username=request.user.username)
-    print(current_user)
+    #print(current_user)
     user_details = ExtraInfo.objects.all().filter(user=current_user).first()
-    print(user_details)
+    #print(user_details)
     user_type = HoldsDesignation.objects.all().filter(user=current_user).first()
-    print(user_type)
+    #print(user_type)
     usertype=str.split(str(user_type))
-    print(usertype)
+    #print(usertype)
     # Academics Admin Check
     user=usertype[0]
     #desig_id = Designation.objects.all().filter(name='Faculty')
     #print(desig_id)
-    print ("yaha se")
-    print(user)
+    #print ("yaha se")
+    #print(user)
     #print(user_type)
     # user = request.user
     # user = User.objects.get(id=1).extrainfo
@@ -911,7 +911,7 @@ def apply_purchase(request):
     # xyz=apply_for_purchase(indentor_name=name,)
     # xyz.save()
 
-        print("hello1")
+        #print("hello1")
 
         a = apply_for_purchase.objects.create(
                 item_name=item_name,
@@ -947,6 +947,54 @@ def apply_purchase(request):
         return HttpResponseRedirect('/office/officeOfPurchaseOfficer/')
 
 
+
+
+
+@login_required
+
+def add_items(request):
+    current_user = get_object_or_404(User, username=request.user.username)
+    print(current_user)
+    user_details = ExtraInfo.objects.all().filter(user=current_user).first()
+    print(user_details)
+    user_type = HoldsDesignation.objects.all().filter(user=current_user).first()
+    print(user_type)
+    usertype=str.split(str(user_type))
+    print(usertype)
+    # Academics Admin Check
+    user=usertype[0]
+    #desig_id = Designation.objects.all().filter(name='Faculty')
+    quantity=0
+    #print(desig_id)
+    print ("yaha se")
+    print(user)
+    if(user == "student"):
+        return HttpResponse('You are not authorised to view this page')
+    if request.method == 'POST':
+        item_name=request.POST.get('item_name')
+        nature_of_item2= 1 if request.POST.get('nature_of_item2') == 'on' else 0
+        quantity=request.POST.get('quant')
+        order_id=request.POST.get('order_id')
+        print(quantity)
+        a = stock.objects.create(
+            item_type=nature_of_item2,
+            item_name=item_name,
+            quantity=quantity,
+        )
+        a.save()
+        messages.info(request, 'item added in inventory successfully')
+        return render(request, "officeModule/officeOfPurchaseOfficer/manageStore_content2.html",{})
+    else:
+        return render(request, "officeModule/officeOfPurchaseOfficer/manageStore_content2.html",{})
+
+
+
+@login_required
+
+def view_items(request):
+    return HttpResponse(1)
+
+
 def submit(request):
     context = {}
 
@@ -971,28 +1019,29 @@ def after_purchase(request):
 def officeOfPurchaseOfficer(request):
     context={}
     current_user = get_object_or_404(User, username=request.user.username)
-    print(current_user)
+    #print(current_user)
     user_details = ExtraInfo.objects.all().filter(user=current_user).first()
-    print(user_details)
+    #print(user_details)
     user_type = HoldsDesignation.objects.all().filter(user=current_user).first()
-    print(user_type)
+    #print(user_type)
     usertype=str.split(str(user_type))
     print(usertype)
     # Academics Admin Check
     user_type=None
     user = usertype[2]
+
     if(user == "student"):
         return HttpResponse('You are not authorised to view this page')
     if(len(usertype)>3):
         user_type=usertype[2]+usertype[3]
     elif(len(usertype)>=2):
         user_type = usertype[2]
-
+    print(user_type)
     p = None
     q = None
     ph = None
-    print ("yaha se")
-    print(user_type)
+    #print ("yaha se")
+    #print(user_type)
     #use user_type to check the designation of actor
     #utype is for rendering tags (templates)
     utype=0#this is for manage store and vendors
@@ -1008,14 +1057,14 @@ def officeOfPurchaseOfficer(request):
     #bug was that pkhanna was being recognized as associate professor not HOD
     if(per_user=="pkhanna" or user_type=="DeputyRegistrar" or user_type=="Registrar" or user_type=="HOD" or user_type=="PurchaseOfficer" or user_type =="Director" or user_type=="director"):
         utype2=1
-        print("hard code working hai yahan tak for HOD designation")
+        #print("hard code working hai yahan tak for HOD designation")
 
-    if(user_type == "AssociateProfessor" or user_type=="HOD" or user_type=="Director" or user_type=="director"):#give more such equivalent designations
+    if(user_type == "AssociateProfessor" or user_type == "AssistantProfessor" or user_type=="HOD" or user_type=="Director" or user_type=="director"):#give more such equivalent designations
         utype3=1
-        print("recognized actor assoc prof")
+        #print("recognized actor assoc prof")
 
 
-    print("yahan tak sab sahi")
+    #print("yahan tak sab sahi")
 
     if request.method == 'POST':
 
@@ -1039,67 +1088,74 @@ def officeOfPurchaseOfficer(request):
             #render appropritate templates as per the actor like approvalHOD2 or approvalRegistrar2 etc
             #this is working through user_type not utype
             if(user_type=="HOD" or per_user=="pkhanna"):
-                alldata = apply_for_purchase.objects.filter(HOD_approve_tag=0)
+                alldata = apply_for_purchase.objects.filter(HOD_approve_tag=0).order_by('-id')
                 context={'alldata':alldata}
                 print(context)
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvalHOD2.html",context=context)
 
             elif(user_type=="DeputyRegistrar" or user_type=="Registrar" or per_user=="swapnali"):
-                alldata = apply_for_purchase.objects.filter(HOD_approve_tag=1,registrar_approve_tag=0)
+                alldata = apply_for_purchase.objects.filter(HOD_approve_tag=1,registrar_approve_tag=0).order_by('-id')
                 #alldata2 = apply_for_purchase.objects.filter(HOD_approve_tag=1,registrar_approve_tag=0,expected_cost__gte = 50001)
-                context = {'alldata':alldata}
+                context = {'alldata':alldata,'des':user_type}
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvalRegistrar.html",context=context)
 
             elif(user_type=="director" or user_type=="Director"):
                 #alldata = apply_for_purchase.objects.filter(HOD_approve_tag=1,registrar_approve_tag=1,expected_cost__gte = 50001)
-                alldata = apply_for_purchase.objects.filter(expected_cost__gte = 50001,director_approve_tag=0)
+                alldata = apply_for_purchase.objects.filter(expected_cost__gte = 50001,director_approve_tag=0).order_by('-id')
                 print(alldata)
-                context = {'alldata':alldata}
+                context = {'alldata':alldata,'des':user_type}
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvaldirector.html",context=context)
 
             elif(user_type=="purchaseofficer" or user_type=="PurchaseOfficer"):
                 print("entered purchase officer section")
-                alldata = apply_for_purchase.objects.filter(director_approve_tag=1,gem_tag=0)
+                alldata = apply_for_purchase.objects.filter(director_approve_tag=1,gem_tag=0).order_by('-id')
                 #alldata2 = apply_for_purchase.objects.filter(HOD_approve_tag=1,registrar_approve_tag=1,expected_cost__lte = 50000)
-                #context = {'alldata':alldata,'alldata2':alldata2}
+                #context = {'alldata':alldata,'alldata2':alldata2}SS
 
-                context = {'alldata':alldata}
-                print("data 0",context)
+                context = {'alldata':alldata,'des':user_type}
+                #print("data 0",context)
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvalpurchaseofficer.html",context=context)
 
 
 
         elif "approved_orders" in request.POST:
             if(user_type=="HOD" or per_user=="pkhanna"):
-                alldata = apply_for_purchase.objects.filter(HOD_approve_tag=1)
+                alldata = apply_for_purchase.objects.filter(HOD_approve_tag=1).order_by('-id')
                 context={'alldata':alldata}
-                print(context)
+                #print(context)
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvedHOD.html",context=context)
 
             elif(user_type=="DeputyRegistrar" or user_type=="Registrar" or per_user=="swapnali"):
-                alldata = apply_for_purchase.objects.filter(registrar_approve_tag=1)
-                context = {'alldata':alldata}
+                alldata = apply_for_purchase.objects.filter(registrar_approve_tag=1,expected_cost__lte = 50000).order_by('-id')
+                context = {'alldata':alldata,'des':user_type}
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvedRegistrar.html",context=context)
 
             elif(user_type=="director" or user_type=="Director"):
                 #alldata = apply_for_purchase.objects.filter(HOD_approve_tag=1,registrar_approve_tag=1,expected_cost__gte = 50001)
-                alldata = apply_for_purchase.objects.filter(expected_cost__gte = 50001,director_approve_tag=1)
-                print(alldata)
+                alldata = apply_for_purchase.objects.filter(expected_cost__gte = 50001,director_approve_tag=1).order_by('-id')
+                #print(alldata)
                 context = {'alldata':alldata}
                 return render(request, "officeModule/officeOfPurchaseOfficer/approvedDirector.html",context=context)
 
-            elif(user_type=="purchaseofficer" or user_type=="PurchaseOfficer"):
-                return HttpResponse("the user template is WUP")
             #make the templates as per the actors
 
-        elif "rejected_orders" in request.POST:
-            return HttpResponse("t")
 
 
 
+        elif "checked_orders" in request.POST:
+            if(user_type == "PurchaseOfficer" or user_type == "purchseofficer"):
+                #print("in checked orders")
+                alldata = apply_for_purchase.objects.filter(director_approve_tag=1).exclude(gem_tag=0)
+                context = {'alldata':alldata,'des':user_type}
+                return render(request, "officeModule/officeOfPurchaseOfficer/checkedpurchaseofficer.html",context=context)
 
+        elif "forwarded_orders" in request.POST:
+            if(user_type == "Registrar" or user_type == "DeputyRegistrar"):
+                #print("in checked orders")
+                alldata = apply_for_purchase.objects.filter(expected_cost__gte = 50001,director_approve_tag=0).order_by('-id')
+                context = {'alldata':alldata,'des':user_type}
 
-
+                return render(request, "officeModule/officeOfPurchaseOfficer/forwardedRegistrar.html",context=context)
 
 
         elif "store" in request.POST:
@@ -1121,6 +1177,14 @@ def officeOfPurchaseOfficer(request):
                 it="non-consumable"
             str2 = "item " + item_name + "of type " + it + " has been added to the store"
             messages.info(request,str2)
+
+        elif "manage_store" in request.POST:
+            return render(request, "officeModule/officeOfPurchaseOfficer/manageStore_content2.html")
+
+        elif "view_store" in request.POST:
+            q=stock.objects.all()
+            return render(request, "officeModule/officeOfPurchaseOfficer/manageStore_content1.html",{"q":q})
+            #return HttpResponse(1)
 
         elif "item_search" in request.POST:
             srch = request.POST['item_name']
@@ -1147,10 +1211,6 @@ def officeOfPurchaseOfficer(request):
             return render(request, "officeModule/officeOfPurchaseOfficer/purchaseHistory_content2.html",{'alldata':alldata})
 
 
-
-
-
-
         elif "purchase_search" in request.POST:
             pr = request.POST['file']
             phmatch = apply_for_purchase.objects.filter(Q(id=pr))
@@ -1168,14 +1228,18 @@ def officeOfPurchaseOfficer(request):
         q=stock.objects.all()
         ph=apply_for_purchase.objects.all()
 
-    return render(request, "officeModule/officeOfPurchaseOfficer/officeOfPurchaseOfficer.html",{'p':p,'q':q,'ph':ph,'utype':utype,'utype2':utype2,'utype3':utype3})
+    return render(request, "officeModule/officeOfPurchaseOfficer/officeOfPurchaseOfficer.html",{'p':p,'q':q,'ph':ph,'utype':utype,'utype2':utype2,'utype3':utype3,'des':user_type})
 
 def delete_item(request,id):
     #template = 'officemodule/officeOfPurchaseOfficer/manageStore_content1.html'
-    print(">>>>>>>")
-    print(id)
-    item = get_object_or_404(stock,id=id)
-    item.delete()
+    print("reached delete_item")
+    if request.method=='POST':
+        objid = request.POST.get('id')
+        print(objid)
+        item = get_object_or_404(stock,id=id)
+        item.delete()
+
+
     return HttpResponse("Deleted successfully")
 
 def delete_vendor(request,id):
